requirements.txt
# Core Django (4.2 LTS - compatível com todas as dependências)
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0

# Celery (Async Tasks)
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Upload & Storage
Pillow==10.1.0
django-storages==1.14.2

# API Documentation
drf-spectacular==0.27.0
django-filter==23.5

# Security
django-encrypted-model-fields==0.6.5
python-decouple==3.8

# Utilities
python-dateutil==2.8.2
pytz==2023.3
qrcode==7.4.2
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.4

# External APIs
requests==2.31.0
httpx==0.25.2

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.18.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.39.1
